+++ src/index.js
@@ -1,14 +1,6 @@
 // Export all constants
 export * from './_constants';
 
-// Export console logger and event observer
-export { default as Logger } from './logger';
-export { default as EventObserver } from './events';
-
-export { unserialize, protobuf } from './protobuf';
-export { default as Protocol } from './types/protocol';
-export * from './types/message';
-
-// Export as default AND as named
+// Export the main class as default AND as named
 export { default as Syft } from './syft';
 export { default } from './syft';
